block_svc_routines: split the code into various new files
split the code from

rpc/
├── block_svc_routines.c

to

rpc/
├── block_common.h
├── block_create.c
├── block_delete.c
├── block_genconfig.c
├── block_info.c
├── block_list.c
├── block_modify.c
├── block_replace.c
├── block_svc_routines.c
├── block_version.c
...

This should help us further organize the code in a better way.
