+++ rpc/Makefile.am
@@ -2,7 +2,10 @@ SUBDIRS = rpcl
 
 noinst_LTLIBRARIES = libgbrpc.la
 
-libgbrpc_la_SOURCES = block_svc_routines.c glfs-operations.c
+libgbrpc_la_SOURCES = block_svc_routines.c block_info.c block_list.c           \
+                      block_create.c block_delete.c block_modify.c             \
+                      block_replace.c block_version.c block_genconfig.c        \
+                      block_common.h glfs-operations.c
 
 noinst_HEADERS = glfs-operations.h
